atomic mock ln rpc in fixtures
