+++ gateway/tests/tests/fixtures/mod.rs
@@ -1,24 +1,19 @@
-use std::sync::{Arc, Mutex};
+use std::sync::Arc;
 
 use anyhow::Result;
-use async_trait::async_trait;
-use bitcoin::{secp256k1, KeyPair};
 use fedimint_api::task::TaskGroup;
-use fedimint_ln::contracts::Preimage;
 use fedimint_testing::btc::{fixtures::FakeBitcoinTest, BitcoinTest};
-use lightning_invoice::Invoice;
 use ln_gateway::{
     client::{GatewayClientBuilder, MemDbFactory},
     config::GatewayConfig,
-    ln::{LightningError, LnRpc},
     rpc::GatewayRequest,
     LnGateway,
 };
-use rand::rngs::OsRng;
 use tokio::sync::mpsc;
 
 pub mod client;
 pub mod fed;
+pub mod ln;
 
 pub struct Fixtures {
     pub bitcoin: Box<dyn BitcoinTest>,
@@ -29,7 +24,7 @@ pub struct Fixtures {
 pub async fn fixtures(gw_cfg: GatewayConfig) -> Result<Fixtures> {
     let task_group = TaskGroup::new();
 
-    let ln_rpc = Arc::new(MockLnRpc::new());
+    let ln_rpc = Arc::new(ln::MockLnRpc::new());
 
     let client_builder: GatewayClientBuilder =
         client::TestGatewayClientBuilder::new(MemDbFactory.into()).into();
@@ -43,41 +38,4 @@ pub async fn fixtures(gw_cfg: GatewayConfig) -> Result<Fixtures> {
         gateway,
         task_group,
     })
-}
-
-struct MockLnRpc {
-    pub preimage: Preimage,
-    node_pubkey: secp256k1::PublicKey,
-    amount_sent: Arc<Mutex<u64>>,
-}
-
-impl MockLnRpc {
-    fn new() -> Self {
-        let ctx = secp256k1::Secp256k1::new();
-        let kp = KeyPair::new(&ctx, &mut OsRng);
-
-        Self {
-            preimage: Preimage([1; 32]),
-            node_pubkey: secp256k1::PublicKey::from_keypair(&kp),
-            amount_sent: Arc::new(Mutex::new(0)),
-        }
-    }
-}
-
-#[async_trait]
-impl LnRpc for MockLnRpc {
-    async fn pubkey(&self) -> Result<secp256k1::PublicKey, LightningError> {
-        Ok(self.node_pubkey)
-    }
-
-    async fn pay(
-        &self,
-        invoice: Invoice,
-        _max_delay: u64,
-        _max_fee_percent: f64,
-    ) -> Result<Preimage, LightningError> {
-        *self.amount_sent.lock().unwrap() += invoice.amount_milli_satoshis().unwrap();
-
-        Ok(self.preimage.clone())
-    }
 }
